+++ PyPWA/libs/file/processor/main.py
@@ -25,7 +25,7 @@ from pathlib import Path
 from typing import Union
 
 import numpy as npy
-import pandas
+import pandas as pd
 
 from PyPWA import info as _info
 from PyPWA.libs.file import cache
@@ -38,7 +38,7 @@ __author__ = _info.AUTHOR
 __version__ = _info.VERSION
 
 
-SUPPORTED_DATA = Union[npy.ndarray, vectors.ParticlePool]
+SUPPORTED_DATA = Union[npy.ndarray, vectors.ParticlePool, pd.DataFrame]
 INPUT_TYPE = Union[Path, str]
 
 
@@ -67,45 +67,45 @@ class _DataLoader:
     __LOGGER = logging.getLogger(__name__ + "._DataLoader")
 
     def __init__(self, use_cache: bool, clear_cache: bool):
-        self.__args = (use_cache, clear_cache)
-        self.__cache_builder = cache.CacheFactory(use_cache, clear_cache)
+        self.__use_cache = use_cache
+        self.__clear_cache = clear_cache
 
     def __repr__(self):
         return (f"{self.__class__.__name__}"
-                f"({self.__args[0]}, {self.__args[1]})")
+                f"({self.__use_cache}, {self.__clear_cache})")
 
-    def parse(self, filename: Path) -> Union[pandas.DataFrame, pandas.Series]:
-        cache_obj = self.__cache_builder.get_cache(filename)
-        if cache_obj.is_valid:
+    def parse(self, filename: Path) -> Union[pd.DataFrame, pd.Series]:
+        valid, cache_obj = cache.read(filename, remove_cache=self.__clear_cache)
+        if valid and self.__use_cache:
             self.__LOGGER.info("Loading cache for %s" % filename)
-            return cache_obj.read_cache()
+            return cache_obj
         else:
             self.__LOGGER.info("No cache found, loading file directly.")
-            return self.__read_data(cache_obj, filename)
+            return self.__read_data(filename)
 
-    @staticmethod
-    def __read_data(cache_obj, filename):
+    def __read_data(self, filename):
         plugin = _get_read_plugin(filename)
         data = plugin.get_memory_parser().parse(filename)
-        cache_obj.write_cache(data)
+        if self.__use_cache:
+            cache.write(filename, data)
         return data
 
 
 class _DataDumper:
 
     def __init__(self, use_cache: bool, clear_cache: bool):
-        self.__args = (use_cache, clear_cache)
-        self.__cache_builder = cache.CacheFactory(use_cache, clear_cache)
+        self.__use_cache = use_cache
+        self.__clear_cache = clear_cache
 
     def __repr__(self):
         return (f"{self.__class__.__name__}"
-                f"({self.__args[0]}, {self.__args[1]})")
+                f"({self.__use_cache}, {self.__clear_cache})")
 
     def write(self, filename: Path, data: SUPPORTED_DATA):
         parser = self.__get_write_plugin(filename, data).get_memory_parser()
-        cache_obj = self.__cache_builder.get_cache(filename)
         parser.write(filename, data)
-        cache_obj.write_cache(data)
+        if self.__use_cache:
+            cache.write(filename, data)
 
     @staticmethod
     def __get_write_plugin(filename: Path,